refactor(admin): consolidate phase 2a helpers
## Overview
Consolidate the follow-up Phase 2a admin cleanup after the initial admin CRUD commit. This removes duplicated controller option builders and service input-normalization helpers, hardens the admin/auth integration, simplifies migration progress writes, and adds the next admin panel shell and information-architecture implementation plan.

## Problem
The first Phase 2a implementation shipped the full admin CRUD surface area, but several patterns were duplicated across the new controllers and services. Department, staff, SLA, team, email, and template option lists were implemented independently in multiple controllers, and nullable string, nullable integer, boolean, and notes normalization logic was repeated across the admin services.

A few follow-up hardening details were also still local: the temporary manual-verification 2FA bypass for the admsupport user needed to be removed, admin routes needed to use the same staff-auth middleware path as the SCP side, Staff::permissions needed to avoid shadowing Spatie's relationship with the legacy column, and migration progress updates could be reduced to a single helper. The next admin shell and IA pass also needed to be captured as a concrete plan before implementation.

## Fix
- add ProvidesModelOptions and move shared department, staff, SLA, team, email, and template option builders into one admin controller trait
- update canned response, department, help topic, staff, and team controllers to use the shared option provider instead of private duplicated methods
- add NormalizesInput and move shared nullable integer, nullable string, trimmed string, and boolean normalization into one admin service trait
- update role, canned response, SLA, department, help topic, filter, email config, staff, and team services to use the shared normalization helpers
- remove the temporary admsupport 2FA bypass from LoginController so normal two-factor flow applies consistently
- switch the admin route group to AuthenticateStaff plus admin.access so admin routes follow the same staff authentication middleware path as SCP routes
- override Staff::getAttribute('permissions') to return Spatie's permissions relationship instead of the legacy permissions column when permission checks resolve the property
- consolidate migration progress status writes through AbstractMigrator::writeProgress
- add the 2026-05-02 admin panel shell and IA implementation plan covering shared isAdmin props, SCP/admin switching, legacy-style admin tabs, and edit-page inner tabs

## Verification
- git diff --cached --check
- php -l on every staged PHP file
